read and write from ipcclient

+++ host/jni/ipcclient.h
@@ -3,6 +3,6 @@
 
 int ipcclient_connect();
 int ipcclient_write(int fd, const void* pBuf, size_t nBytes);
-int ipcclient_read(int fd, void* pBuf, size_t nBytes);
+size_t ipcclient_read(int fd, void* pBuf, size_t nBytes);
 
 #endif // IPCCLIENT_H_
